Validate queries before actually executing them.
The bot now takes advantage of WolframAlpha's ability to validate the
syntax of a query before executing it. This allows for the bot to more
quickly alert the user if their query would not be understood, since
wa queries usually have a lagtime of several seconds before returning
output. This way, we avoid waiting more than we have to.